Key generation fixes

+++ src/services/adminService.js
@@ -58,9 +58,9 @@ const createOrganization = async (req, res) => {
             //create default provider
             await adminDao.createProvider(organization.ORG_ID, { name: 'WSO2', providerURL: config.controlPlane.url }, t);
             //store default subscription policies
-            const storagePlans  = constants.DEFAULT_SUBSCRIPTION_PLANS;
+            const storagePlans = constants.DEFAULT_SUBSCRIPTION_PLANS;
             for (const plan of storagePlans) {
-              await apiDao.createSubscriptionPolicy(orgId, plan, t);
+                await apiDao.createSubscriptionPolicy(orgId, plan, t);
             }
         });
         const orgCreationResponse = {
@@ -817,6 +817,7 @@ const createAppKeyMapping = async (req, res) => {
             // add subscription to control plane for each api
             const apiSubscriptions = [];
             for (const api of apis) {
+                console.log("API", api);
                 const policyDetails = await apiDao.getSubscriptionPolicy(api.policyID, orgID, t);
                 const cpSubscribeResponse = await createCPSubscription(req, api.apiRefId, cpAppID, policyDetails);
                 apiSubscriptions.push(cpSubscribeResponse);
@@ -839,8 +840,13 @@ const createAppKeyMapping = async (req, res) => {
                     await adminDao.createApplicationKeyMapping(appKeyMappping, t);
                 }
             }
+            //check if more than one mapping en
+            const appKeyMapping = await adminDao.getApplicationKeyMapping(orgID, appID, true);
             //delete app key mapping entries with no api id ref
-            await adminDao.deleteAppKeyMapping(orgID, appID, null, t);
+            if (appKeyMapping.length > 1) {
+                await adminDao.deleteAppKeyMapping(orgID, appID, null, t);
+            }
+            tokenDetails.additionalProperties = checkAdditionalValues(tokenDetails.additionalProperties);
             //generate oauth key
             responseData = await invokeApiRequest(req, 'POST', `${controlPlaneUrl}/applications/${cpAppID}/generate-keys`, {}, tokenDetails);
         });
@@ -849,6 +855,19 @@ const createAppKeyMapping = async (req, res) => {
         console.error(`${constants.ERROR_MESSAGE.KEY_MAPPING_CREATE_ERROR}`, error);
         return util.handleError(res, error);
     }
+}
+
+function checkAdditionalValues(additionalValues) {
+
+    let defaultConfigs = ["application_access_token_expiry_time", "user_access_token_expiry_time", "id_token_expiry_time", "refresh_token_expiry_time"];
+    const props = {}
+    for (const key in additionalValues) {
+        if (defaultConfigs.includes(key)) {
+            props[key] = parseInt(additionalValues[key]);
+        }
+    }
+    return props;
+
 }
 
 const createCPApplication = async (req, cpApplicationName) => {
@@ -1030,5 +1049,6 @@ module.exports = {
     unsubscribeAPI,
     createAppKeyMapping,
     retriveAppKeyMappings,
-    getApplicationKeyMap
+    getApplicationKeyMap,
+    checkAdditionalValues
 };
