+++ src/services/apiMetadataService.js
@@ -19,6 +19,7 @@
 const { Sequelize } = require("sequelize");
 const sequelize = require("../db/sequelize");
 const apiDao = require("../dao/apiMetadata");
+const adminDao = require("../dao/admin");
 const util = require("../utils/util");
 const path = require("path");
 const fs = require("fs").promises;
@@ -300,6 +301,11 @@ const deleteAPIMetadata = async (req, res) => {
     }
     await sequelize.transaction(async (t) => {
         try {
+            //check if subscriptions exist for the application
+            const subApis = await adminDao.getSubscriptions(orgId, '', apiId);
+            if (subApis.length > 0) {
+                throw new CustomError(401, constants.ERROR_CODE[401], "API has subscriptions.");
+            }
             const apiDeleteResponse = await apiDao.deleteAPIMetadata(orgId, apiId, t);
             if (apiDeleteResponse === 0) {
                 throw new Sequelize.EmptyResultError("Resource not found to delete");
@@ -317,7 +323,6 @@ const createAPITemplate = async (req, res) => {
 
     try {
         const { orgId, apiId } = req.params;
-        let imageMetadata = JSON.parse(req.body.imageMetadata);
         const extractPath = path.join("/tmp", orgId + "/" + apiId);
         await fs.mkdir(extractPath, { recursive: true });
         const zipFilePath = req.file.path;
@@ -335,8 +340,12 @@ const createAPITemplate = async (req, res) => {
 
         // Verify directories exist
         try {
-            await fs.access(contentPath);
-            await fs.access(imagesPath);
+            if (fsDir.existsSync(contentPath)) {
+                await fs.access(contentPath);
+            }
+            if (fsDir.existsSync(imagesPath)) {
+                await fs.access(imagesPath);
+            }
             if (fsDir.existsSync(documentPath)) {
                 await fs.access(documentPath);
             }
@@ -348,10 +357,18 @@ const createAPITemplate = async (req, res) => {
                 , Documents path: ${documentPath}`
             );
         }
+        let apiContent = [];
+
         //get api files
-        let apiContent = await util.getAPIFileContent(contentPath);
+        if (fsDir.existsSync(contentPath)) {
+            let apiLanding = await util.getAPIFileContent(contentPath);
+            apiContent.push(...apiLanding);
+        }
         //get api images
-        const apiImages = await util.getAPIImages(imagesPath);
+        if (fsDir.existsSync(imagesPath)) {
+            const apiImages = await util.getAPIImages(imagesPath);
+            apiContent.push(...apiImages);
+        }
         //get api documents
         if (fsDir.existsSync(documentPath)) {
             const apiDocuments = await util.readDocFiles(documentPath);
@@ -363,7 +380,10 @@ const createAPITemplate = async (req, res) => {
             const links = util.getAPIDocLinks(docMetadata);
             apiContent.push(...links);
         }
-        apiContent.push(...apiImages);
+        let imageMetadata = {};
+        if (req.body.imageMetadata) {
+            imageMetadata = JSON.parse(req.body.imageMetadata);
+        }
         await sequelize.transaction(async (t) => {
             //check whether api belongs to given org
             let apiMetadata = await apiDao.getAPIMetadata(orgId, apiId, t);
@@ -372,7 +392,7 @@ const createAPITemplate = async (req, res) => {
             if (imageMetadata[constants.API_ICON], exsistingAPIImage) {
                 await apiDao.deleteImage(constants.API_ICON, apiId, t);
             }
-            
+
             if (apiMetadata) {
                 // Store image metadata
                 await apiDao.storeAPIImageMetadata(imageMetadata, apiId, t);
@@ -393,7 +413,9 @@ const updateAPITemplate = async (req, res) => {
 
     try {
         const { orgId, apiId } = req.params;
-        const imageMetadata = JSON.parse(req.body.imageMetadata);
+        if (req.body.imageMetadata) {
+            imageMetadata = JSON.parse(req.body.imageMetadata);
+        }
         const extractPath = path.join("/tmp", orgId + "/" + apiId);
         await fs.mkdir(extractPath, { recursive: true });
         const zipFilePath = req.file.path;
@@ -408,9 +430,15 @@ const updateAPITemplate = async (req, res) => {
 
         // Verify directories exist
         try {
-            await fs.access(contentPath);
-            await fs.access(imagesPath);
-            await fs.access(documentPath);
+            if (fsDir.existsSync(contentPath)) {
+                await fs.access(contentPath);
+            }
+            if (fsDir.existsSync(imagesPath)) {
+                await fs.access(imagesPath);
+            }
+            if (fsDir.existsSync(documentPath)) {
+                await fs.access(documentPath);
+            }
         } catch (err) {
             console.error(err);
             throw new Error(
@@ -418,20 +446,28 @@ const updateAPITemplate = async (req, res) => {
                 Documents path: ${documentPath}`
             );
         }
+        let apiContent = [];
         //get api files
-        let apiContent = await util.getAPIFileContent(contentPath);
+        if (fsDir.existsSync(contentPath)) {
+            let apiLanding = await util.getAPIFileContent(contentPath);
+            apiContent.push(...apiLanding);
+        }
         //get api images
-        const apiImages = await util.getAPIImages(imagesPath);
+        if (fsDir.existsSync(imagesPath)) {
+            const apiImages = await util.getAPIImages(imagesPath);
+            apiContent.push(...apiImages);
+        }
         //get api documents
-        const apiDocuments = await util.readDocFiles(documentPath);
+        if (fsDir.existsSync(documentPath)) {
+            const apiDocuments = await util.readDocFiles(documentPath);
+            apiContent.push(...apiDocuments);
+        }
 
         if (req.body.docMetadata) {
             docMetadata = JSON.parse(req.body.docMetadata);
             const links = util.getAPIDocLinks(docMetadata);
             apiContent.push(...links);
         }
-        apiContent.push(...apiDocuments);
-        apiContent.push(...apiImages);
         await sequelize.transaction(async (t) => {
             //check whether api belongs to given org
             const apiMetadata = await apiDao.getAPIMetadata(orgId, apiId, t);
